fix: 기능 권한 제한 모두 추가 (DELETE) add: 마이페이지 조회 기능 머지

+++ server/src/main/java/com/splashzone/boardclub/controller/BoardClubController.java
@@ -96,7 +96,12 @@ public class BoardClubController {
     }
 
     @DeleteMapping("/{club-id}")
-    public ResponseEntity deleteBoardClub(@PathVariable("club-id") @Positive Long boardClubId) {
+    public ResponseEntity deleteBoardClub(Authentication authentication,
+                                          @PathVariable("club-id") @Positive Long boardClubId) {
+        if (authentication == null) {
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
+
         boardClubService.deleteBoardClub(boardClubId);
 
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
